requirements.txt
requests==2.31.0 
beautifulsoup4==4.12.2 
selenium==4.11.2
python-dotenv==1.0.0 
google-generativeai==0.3.1 
linkedin-api==2.0.0a6